# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/certificates": {
      "get": {
        "operationId": "CertificatesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "trackId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "own",
                "any"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เกียรติบัตรของผู้เรียก (เมนูความสำเร็จ) · scope=any สำหรับผู้ดูแล",
        "tags": [
          "certificates"
        ]
      }
    },
    "/api/v1/certificates/{id}": {
      "get": {
        "operationId": "CertificatesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ข้อมูลเกียรติบัตรสำหรับแสดงและดาวน์โหลด",
        "tags": [
          "certificates"
        ]
      }
    },
    "/api/v1/certificate-templates": {
      "get": {
        "operationId": "CertificateTemplatesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เทมเพลตเกียรติบัตร",
        "tags": [
          "certificate-templates"
        ]
      },
      "post": {
        "operationId": "CertificateTemplatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCertificateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เพิ่มเทมเพลตเกียรติบัตร",
        "tags": [
          "certificate-templates"
        ]
      }
    },
    "/api/v1/certificate-templates/{id}": {
      "get": {
        "operationId": "CertificateTemplatesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "certificate-templates"
        ]
      },
      "patch": {
        "operationId": "CertificateTemplatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCertificateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "certificate-templates"
        ]
      },
      "delete": {
        "operationId": "CertificateTemplatesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "certificate-templates"
        ]
      }
    },
...
    "/api/v1/instructor-assignments": {
      "get": {
        "operationId": "AssignmentsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "trackId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "อาจารย์ประจำสายงาน/วิชา (อาจารย์เห็นเฉพาะของตัวเอง)",
        "tags": [
          "instructor-assignments"
        ]
      },
      "post": {
        "operationId": "AssignmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "มอบหมายอาจารย์ให้ดูแลสายงานหรือวิชา",
        "tags": [
          "instructor-assignments"
        ]
      }
    },
    "/api/v1/instructor-assignments/{id}": {
      "delete": {
        "operationId": "AssignmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ยกเลิกการมอบหมาย",
        "tags": [
          "instructor-assignments"
        ]
      }
    }
  },
  "info": {
    "title": "CSMJU E-Learning API",
    "description": "เรียนตามสายงานวิทยาการคอมพิวเตอร์ · token มาจาก Core Hub SSO",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCertificateTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "เทมเพลตมาตรฐาน"
          },
          "heading": {
            "type": "string",
            "example": "เกียรติบัตร"
          },
          "bodyText": {
            "type": "string",
            "description": "ใช้ {name} {track} {date} {certificateNo} แทนค่าได้"
          },
          "signerName": {
            "type": "string"
          },
          "signerTitle": {
            "type": "string"
          },
          "imageId": {
            "type": "object",
            "description": "id ของรูปที่อัปโหลดไว้ใน Core Hub (POST /api/v1/images)"
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "heading",
          "bodyText",
          "signerName",
          "signerTitle"
        ]
      },
      "UpdateCertificateTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "เทมเพลตมาตรฐาน"
          },
          "heading": {
            "type": "string",
            "example": "เกียรติบัตร"
          },
          "bodyText": {
            "type": "string",
            "description": "ใช้ {name} {track} {date} {certificateNo} แทนค่าได้"
          },
          "signerName": {
            "type": "string"
          },
          "signerTitle": {
            "type": "string"
          },
          "imageId": {
            "type": "object",
            "description": "id ของรูปที่อัปโหลดไว้ใน Core Hub (POST /api/v1/images)"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
      "CreateAssignmentDto": {
        "type": "object",
        "properties": {
          "personCode": {
            "type": "string",
            "description": "รหัสบุคลากรใน Core Hub (personCode)"
          },
          "trackId": {
            "type": "string",
            "description": "อาจารย์ประจำสายงาน (ดูแลทุกวิชาในสายงาน)"
          },
          "courseId": {
            "type": "string",
            "description": "ผู้รับผิดชอบวิชาเดียว"
          }
        },
        "required": [
          "personCode"
        ]
      }
    }
  }
}